Begun the long process of upgrading the game to use HD graphics and 16:9 aspect ratio. Many things will be out of place, but functional nonetheless. Exposed missing Size metrics for many graphic elements (still a work in progress) Upgraded and reenabled the Resolution option in the Options Screen. Upgraded Menu objects to handle sound events themselves. No more menu sounds in inappropriate places! Added MENU_INVALID_DECIDE sound event (handled by Menus)

+++ WindowsGame1/WindowsGame1/Drawing/CoopLifebar.cs
@@ -159,7 +159,7 @@ namespace WGiBeat.Drawing
 
             int[] blockAssignments = AssignBlocks(beatFraction);
 
-            int posX = this.X + 3;
+            float posX = this.X + 3;
 
             for (int y = 0; y < blockAssignments.Length; y++)
             {
